+++ src/typescript/applications/rollModifierDialog.ts
@@ -59,6 +59,12 @@ export default class RollModifierDialog extends Application<Options> {
       if (isNaN(number)) {
         throw `The value of the input is not a number: ${input.value}`;
       }
+      if (
+        (this.options.min && number < this.options.min) ||
+        (this.options.max && number > this.options.max)
+      ) {
+        throw `The value of the input is out of bounds: ${input.value}`;
+      }
       this.callback(number);
       this.close();
     } else {