+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- back-end bounds validation for the roll modifier dialog
+
 ## [0.3.1] - 2021-07-17
 
 ### Added